Remove Columnar::Ref and replace by type defs
The Columnar::Ref type needs to be equivalent to the container's reference
type. Remove the duplication and make life easy by introducing a type
definition of the same name that navigates from a type `T` to the reference
type of the columnar container.

Similarly, introduces the `ContainerOf` type definition that takes a
columnar type `T` and navigates to the columnar container's type.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>